Add description in about Header
