+++ _about/Header/data.json
@@ -1,3 +1,4 @@
 {
-    "title": "안녕하세요, 저는 이강혁입니다."
+    "title": "안녕하세요, 저는 이강혁입니다.",
+    "description": "1년차 주니어 프론트엔드 개발자로 분야에 관계 없이 다양한 일들을 경험해보는 것을 좋아합니다.<br />안주하는 것보다는 경험을 통해 변화하고 성장하는 것에 가치가 있다고 느낍니다."
 }
